+++ crates/module-system/sov-modules-api/src/rest/mod.rs
@@ -36,6 +36,7 @@ use sov_rest_utils::Query;
 use crate::hooks::TxHooks;
 use crate::map::NamespacedStateMap;
 use crate::rest::__private::maybe_archival_accessor;
+pub use crate::rest::__private::openapi::OpenApi;
 use crate::vec::NamespacedStateVec;
 use crate::{ApiStateAccessor, Module, ModuleId, ModuleInfo, Spec};
 
@@ -160,7 +161,7 @@ pub trait HasCustomRestApi: Sized + Clone {
 
     /// Returns the OpenAPI specification for [`HasCustomRestApi::custom_rest_api`].
     /// [`None`] means there is no known OpenAPI spec for the API.
-    fn custom_openapi_spec(&self) -> Option<serde_json::Value> {
+    fn custom_openapi_spec(&self) -> Option<OpenApi> {
         None
     }
 }