+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 - #1370 Makes some light changes to the `GenesisStateAccessor` to become a wrapper around `StateCheckpoint`. 
 - #1362 Adds a versioned state accessor to be used in the soft confirmation context. This versioned state accessor is append-only and should be initialized at genesis to be properly used.
 - #1366 Replaces the `Delta` by a `StateCheckpoint` inside the `TxScratchpad`. This is because we are going to add fields to the `StateCheckpoint` (`virtual_height`, `true_height`) - this will allow to propagate the values up to the `WorkingSet`.
+- #1356 Adds OpenAPI spec for sov-bank custom REST API endpoints.
 - #1374 Add gas handling in the stf-blueprint::process_proof.
 - #1377 Enables REST API endpoints for `AttesterIncentives` module.
 - #1369 add an extra generic type to the `sov_evm::authenticate` function. This generic enables conversion between `reth::address` and `Spec::Address`. This change is only breaking for users of the EVM module.